requirements.txt
Django==1.11.3
django-orientdb-api==0.1
httplib2==0.10.3
pytz==2017.2
